remove unnecessary package in code and create a new netcomponent
