+++ Assets/IRXRClient/Scripts/NetComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ebdb647b98aa6f16ab352d70b3d9300
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
