+++ Assets/SceneLoader/Scripts/PointCloudLoader.cs
@@ -1,8 +1,6 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using Newtonsoft.Json;
+
 
 class PointCloudData {
     public float[] positions;
